+++ resources/__init__.py
@@ -1,0 +1,3 @@
+"""
+Pi-hole MCP resources package
+"""